+++ plugins/periskop-backend/src/plugin.ts
@@ -31,7 +31,7 @@ export const periskopPlugin = createBackendPlugin({
   register(env) {
     env.registerInit({
       deps: {
-        config: coreServices.config,
+        config: coreServices.rootConfig,
         logger: coreServices.logger,
         httpRouter: coreServices.httpRouter,
       },