+++ plugins/kafka-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-kafka-backend
 
+## 0.2.41-next.1
+
+### Patch Changes
+
+- 629cbd194a87: Use `coreServices.rootConfig` instead of `coreService.config`
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.1
+  - @backstage/backend-plugin-api@0.6.0-next.1
+  - @backstage/catalog-model@1.4.1
+  - @backstage/config@1.0.8
+  - @backstage/errors@1.2.1
+
 ## 0.2.41-next.0
 
 ### Patch Changes