+++ plugins/periskop-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-periskop-backend
 
+## 0.1.19-next.1
+
+### Patch Changes
+
+- 629cbd194a87: Use `coreServices.rootConfig` instead of `coreService.config`
+- 12a8c94eda8d: Add package repository and homepage metadata
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.1
+  - @backstage/backend-plugin-api@0.6.0-next.1
+  - @backstage/config@1.0.8
+
 ## 0.1.19-next.0
 
 ### Patch Changes